% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{get_correct_answers,quiz-method}
\alias{get_correct_answers,quiz-method}
\title{Get correct answers from a quiz object}
\usage{
\S4method{get_correct_answers}{quiz}(quiz)
}
\arguments{
\item{quiz}{The quiz object to access}
}
\value{
Integer representing the number of correct answers
}
\description{
This method retrieves the number of correct answers from a quiz.
}
\examples{
get_correct_answers(myQuiz)
}
