% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ui.R
\name{get_image_base64}
\alias{get_image_base64}
\title{Convert Image to Base64 URL}
\usage{
get_image_base64(image_name, package_name = "vokabel")
}
\arguments{
\item{image_name}{The name of the image file to be converted to a base64 URL.}

\item{package_name}{The name of the package where the image is stored, with a default value of "vokabel".}
}
\value{
A string containing the base64 URL of the image.
}
\description{
This function reads an image file from the package's \code{www} directory and converts it to a base64 URL using the magick package.
It allows for dynamic selection of images stored within the package, making it versatile for use in web applications, especially Shiny apps.
}
\examples{
get_image_base64("logo.png")
get_image_base64("abc.png", package_name = "vokabel")
}
