% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-module.R
\name{get_quiz_stats}
\alias{get_quiz_stats}
\title{Get correct answers and total questions}
\usage{
get_quiz_stats(quiz)
}
\arguments{
\item{quiz}{The quiz object}
}
\value{
A named list containing the number of correct answers and total questions
}
\description{
This function returns the current number of correct answers and total questions.
}
