% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{get_total_questions,quiz-method}
\alias{get_total_questions,quiz-method}
\title{Get total questions from a quiz object}
\usage{
\S4method{get_total_questions}{quiz}(quiz)
}
\arguments{
\item{quiz}{The quiz object to access}
}
\value{
Integer representing the total number of questions
}
\description{
This method retrieves the total number of questions in a quiz.
}
\examples{
get_total_questions(myQuiz)
}
